+++ api/animals/coordinators/animal_coordinator.py
@@ -8,7 +8,9 @@ from ..data_access.animal_status_provider import AnimalStatusProvider
 from ..data_access.animal_viewable_on_day_provider import AnimalViewableOnDayProvider
 from ..data_access.animal_viewing_alert_provider import AnimalViewingAlertProvider
 from ..data_access.animal_viewing_scope_provider import AnimalViewingScopeProvider
+from ..data_access.animal_visibility_schedule_exhibit_name_provider import AnimalVisibilityScheduleExhibitNameProvider
 from ..data_access.animal_visibility_schedule_provider import AnimalVisibilityScheduleProvider
+from ..data_access.animal_visibility_schedule_species_name_provider import AnimalVisibilityScheduleSpeciesNameProvider
 from ..domain.animal_viewability_builder import AnimalViewabilityBuilder
 from ..domain.animal_viewability_context_builder import AnimalViewabilityContextBuilder
 from ..domain.itinerary_animal_records_filter_builder import ItineraryAnimalRecordsFilterBuilder
@@ -95,6 +97,26 @@ class AnimalCoordinator():
    @classmethod
    def get_off_display_exhibit_options( cls ) -> list[ str ]:
       return AnimalOffDisplayExhibitNameProvider.fetch_off_display_exhibit_names(
+         RequestConnectionProvider.get(),
+         today=DateValues.today_date_key() )
+
+
+   @classmethod
+   def get_animal_visibility_schedule_options( cls, exhibit: str | None = None ) -> list[ str ]:
+      if exhibit:
+         return AnimalVisibilityScheduleSpeciesNameProvider.fetch_visibility_schedule_species_names_in_exhibit(
+            RequestConnectionProvider.get(),
+            today=DateValues.today_date_key(),
+            exhibit=exhibit )
+
+      return AnimalVisibilityScheduleSpeciesNameProvider.fetch_visibility_schedule_species_names(
+         RequestConnectionProvider.get(),
+         today=DateValues.today_date_key() )
+
+
+   @classmethod
+   def get_animal_visibility_schedule_exhibit_options( cls ) -> list[ str ]:
+      return AnimalVisibilityScheduleExhibitNameProvider.fetch_visibility_schedule_exhibit_names(
          RequestConnectionProvider.get(),
          today=DateValues.today_date_key() )