+++ tests/scripts/consoleOperations/animals/controllers/removeVisibilityController.test.mjs
@@ -25,7 +25,7 @@ test('Test_CreateRemoveVisibilityScheduleController_TestShowAndSubmitSuccess_Exp
 
    ControllerHelper.loadOptionsAndShowPanel = async (options) => {
       activations.push(options.panelEl);
-      assert.equal(options.loadOptions, ConsoleOptionsLoader.loadExhibits);
+      assert.equal(options.loadOptions, ConsoleOptionsLoader.loadVisibilityScheduleExhibits);
       assert.equal(options.populateOptions, ConsoleDropdownPopulator.populateExhibitDropdown);
    };
    ConsoleStatusPresenter.setStatus = (...args) => {